// src/modules/operating-room/surgery-patient-safety-checklist/models/safety-checklist.model.ts
export interface ISafetyChecklistForm {
    Waktu: string;
    ID_Petugas: string;
    Updated_At: string;
    Updated_By: string;
    Nama_Petugas: string;
    Sign_In_Tanda: number;
    Sign_In_Waktu: string;
    Sign_In_Alergi: number;
    Sign_In_Implan: number;
    Sign_Out_Waktu: string;
    Time_Out_Hasil: number;
    Time_Out_Waktu: string;
    Sign_In_Lengkap: number;
    Time_Out_Steril: number;
    Sign_In_Informed: number;
    Sign_Out_ID_Mata: string;
    Time_Out_Masalah: string;
    Sign_In_ID_Dokter: string;
    Sign_In_ID_Penata: string;
    Sign_Out_TTD_Mata: string;
    Sign_In_ID_Perawat: string;
    Sign_In_Pendarahan: number;
    Sign_In_Pernafasan: number;
    Sign_In_TTD_Dokter: string;
    Sign_In_TTD_Penata: string;
    Sign_Out_ID_Dokter: string;
    Sign_Out_ID_Penata: string;
    Sign_Out_Nama_Mata: string;
    Time_Out_ID_Dokter: string;
    Time_Out_ID_Penata: string;
    Time_Out_Peralatan: number;
    Sign_In_Nama_Dokter: string;
    Sign_In_Nama_Penata: string;
    Sign_In_TTD_Perawat: string;
    Sign_Out_ID_Perawat: string;
    Sign_Out_TTD_Dokter: string;
    Sign_Out_TTD_Penata: string;
    Time_Out_Baca_Ulang: number;
    Time_Out_ID_Perawat: string;
    Time_Out_Pendarahan: number;
    Time_Out_TTD_Dokter: string;
    Time_Out_TTD_Penata: string;
    Sign_In_Nama_Perawat: string;
    Sign_Out_ID_Sirkuler: string;
    Sign_Out_Nama_Dokter: string;
    Sign_Out_Nama_Penata: string;
    Sign_Out_TTD_Perawat: string;
    Time_Out_Nama_Dokter: string;
    Time_Out_Nama_Penata: string;
    Time_Out_TTD_Perawat: string;
    Time_Out_Tidak_Rutin: number;
    Sign_Out_Nama_Perawat: string;
    Sign_Out_TTD_Sirkuler: string;
    Time_Out_Nama_Perawat: string;
    Sign_In_Ruangan_Dokter: number;
    Sign_In_Ruangan_Penata: number;
    Sign_Out_Nama_Sirkuler: string;
    Sign_Out_Nama_Tindakan: number;
    Sign_Out_Ruangan_Bedah: number;
    Time_Out_Ruangan_Bedah: number;
    Sign_In_Pulse_Oksimetri: number;
    Sign_In_Ruangan_Perawat: number;
    Sign_Out_Catatan_Khusus: number;
    Sign_Out_Ruangan_Dokter: number;
    Sign_Out_Ruangan_Penata: number;
    Time_Out_Ruangan_Dokter: number;
    Time_Out_Ruangan_Penata: number;
    Sign_Out_Ruangan_Perawat: number;
    Time_Out_Anestesi_Khusus: number;
    Time_Out_Perkenalan_Diri: number;
    Time_Out_Ruangan_Perawat: number;
    Sign_Out_Kelengkapan_Alat: number;
    Sign_Out_Masalah_Peralatan: number;
    Time_Out_Tidak_Rutin_Waktu: string;
    Sign_Out_Pelabelan_Spesimen: number;
    Time_Out_Profilaksis_Antibiotik: number;
}
// ...
export interface ISafetyChecklist extends IDataModel {
    form: ISafetyChecklistForm;
}

export class SafetyChecklist extends DataModel {
    form: ISafetyChecklistForm;
    constructor(req: ISafetyChecklist) {
      super(req)
      this.form = req.form;
    }
}